requirements.txt
langchain-community
langchain
langchain-core
langchain-community
python-dotenv
langchain_openai
pymupdf
docarray
pinecone-client
tqdm